Crate arena

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The arena, a fast but limited type of allocator.

Arenas are a type of allocator that destroy the objects within, all at once, once the arena itself is destroyed. They do not support deallocation of individual objects while the arena itself is still alive. The benefit of an arena is very fast allocation; just a pointer bump.

This crate implements TypedArena, a simple arena that can only hold objects of a single type.

Structs

DroplessArena 	 [
Experimental
]

TypedArena 	 [
Experimental
]
An arena that can hold objects of only one type.