Struct arena::DroplessArena

pub struct DroplessArena { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl DroplessArena

pub fn new() -> DroplessArena

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn in_arena<T: ?Sized>(&self, ptr: *const T) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn alloc<T>(&self, object: T) -> &mut T

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn alloc_slice<T>(&self, slice: &[T]) -> &mut [T] where
T: Copy,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Allocates a slice of objects that are copied into the DroplessArena, returning a mutable reference to it. Will panic if passed a zero-sized type.

Panics: - Zero-sized types - Zero-length slices