Crate arena

Unstable (rustc_private)

The arena, a fast but limited type of allocator.

Arenas are a type of allocator that destroy the objects within, all at once, once the arena itself is destroyed. They do not support deallocation of individual objects while the arena itself is still alive. The benefit of an arena is very fast allocation; just a pointer bump.

This crate has two arenas implemented: TypedArena, which is a simpler arena but can only hold objects of a single type, and Arena, which is a more complex, slower arena which can hold objects of any type.

Structs

TypedArena 	 [Unstable]
A faster arena that can hold objects of only one type.